sign conversions fixed

+++ testing/tst_array.h
@@ -13,12 +13,12 @@ TEST(array, access) {
     array<int, 10> a;
 
     for (size_t i = 0; i < a.size(); i++) {
-        a[i] = a.size() - i;
+        a[i] = static_cast<int>(a.size() - i);
         EXPECT_EQ(size_t(a.size()) - i, a[i]);
     }
 
     for (size_t i = 0; i < a.size(); i++) {
-        a.at(i) = i;
+        a.at(i) = static_cast<int>(i);
         EXPECT_EQ(i, a[i]);
     }
 
@@ -28,7 +28,7 @@ TEST(array, access) {
 TEST(array, copy) {
     array<int, 30> a;
     for (size_t i = 0; i < a.size(); i++) {
-        a[i] = a.size() - i;
+        a[i] = static_cast<int>(a.size() - i);
         EXPECT_EQ(size_t(a.size()) - i, a[i]);
     }
